Resolve importing macro issued
The previous commit had the problem that we needed to explicitly
import the "private" helper function log::timestamp(), because it
needs to have the same scope as the macro calling it. Now we got
around that extra line of code by calling this function with its
complete path (i.e., `crate::log::timestamp()`).