I can't believe I just wrote a macro for a simple table
I can't get my head around two conflicting demands: I want to have
constant values which are defined at runtime, and yet I can't have a
"const" because we will have standard headers as well as custom
headers, and thus need a struct which owns the string it contains.

After fighting the borrow checker, I finally decided to just use an
enum.

It's funny that such an operation is so easy in Java, where you can
associate an enum with a specific value by passing it to the
constructor. Writing a macro for a simple map looks, well, a bit like
overengineering.